requirements.txt
# Core dependencies - Updated for Python 3.11.9
pywin32>=306
pythonnet>=3.0.3
psutil>=7.1.2
pydantic>=2.5.2
loguru>=0.7.2
pyyaml>=6.0.1
watchdog>=3.0.0

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Build dependencies
pyinstaller>=6.16.0
auto-py-to-exe>=2.40.0

# Optional: For advanced features
asyncio-mqtt>=0.16.1
redis>=5.0.1
